Add `list car` command and refactor CAR admin server handlers
Add a CLI command that lists the imported CAR paths. The command is
added to `list` as `car` sub-command. As a result, the listing of
advertisements command is now moved to the `list ad` sub-command.

Combine the CAR handlers on admin server for simplicity.

Fixes: #119
